+++ src/cli/commands/private-cmds/ci-update-cmd.js
@@ -8,11 +8,14 @@ export default class CiUpdate extends Command {
   name = 'ci-update <id> [scopePath]';
   description = 'run an update for build and test of a certain bit-component';
   alias = '';
-  opts = [];
+  opts = [
+    ['v', 'verbose', 'showing npm verbose output for inspection'],
+  ];
   private = true;
 
-  action([id, scopePath, ]: [string, ?string]): Promise<any> {
-    return ciUpdateAction(id, scopePath || process.cwd());
+  action([id, scopePath]: [string, ?string, ], { verbose }: { verbose: ?boolean }): Promise<any> {
+    verbose = true; // During ci-update we always want to see verbose outputs
+    return ciUpdateAction(id, scopePath || process.cwd(), verbose);
   }
 
   report(maybeSpecsResults: SpecsResults|Error): string {