+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - added `--environment` option for `bit import` 
 - reformatted `bit import` output (components, dependencies, environments)
 - remove duplication for missing packages warning
+- Remove the npm tree output for component ci flow
+- add verbosity option to some places
+
 
 ## [0.10.6] - 2017-08-23